Changed paddle to cylinder with sphere ends
To make the paddle more round looking, I changed it to a horizontal cylinder (size 5) with two sphere's on each end to round it off.

Also added paddleSpeed float as a public variable that can be changed without going into a text editor.

Edited Ball Script so it registers when it collides with 'Paddle' instead of 'paddle'. This was to keep with all prefabs named with a capitol letter.