Attempted to remove counting bug
changed counter to a list and used .RemoveAll(item => item == null) function to have the list check it's items for any null values and to remove them if found. This should be faster than checking for all the bricks on screen and seems to be about as effective.
Level 4 seems to do weird things to the counter will continue fixing bug.