Academy added. Reworking scripts to get all bricks in paddleAgent
Academy is now all added but when training it gives up after the first iteration. I think this is due to a null reference error when the bricks reinstantiate themselves. Still working at getting the bricks to be found and instantiated for each of the agents and not all at once.